Split every filter and synth in their own file
