+++ music/__init__.py
@@ -1,2 +1,8 @@
-from . import utils, tables, synths, effects, structures, singing, core
+from . import core
+from . import effects
+from . import singing
+from . import structures
+from . import tables
+from . import utils
+
 from . import legacy
