+++ music/core/__init__.py
@@ -1,4 +1,8 @@
-from .functions import *
 from .classes import *
+from .filters import *
 
+# from .functions import *
+# from .io import *
+# from .synths import *
 
+# from music.utils import *
