+++ music/core/filters/__init__.py
@@ -1,0 +1,25 @@
+__all__ = [
+    "ad",
+    "ads",
+    "adv",
+    "am",
+    "cf",
+    "f",
+    "fir",
+    "iir",
+    "l",
+    "l_",
+    "r"
+]
+
+from .ad import AD
+from .ads import ADS
+from .adv import ADV
+from .am import AM
+from .cf import CF
+from .f import F
+from .fir import FIR
+from .iir import IIR
+from .l import L
+from .l_ import L_
+from .r import R
