+++ music/core/filters/adv.py
@@ -1,0 +1,5 @@
+from .ad import AD
+from ..synths.v_ import V_
+
+def ADV(note_dict={}, adsr_dict={}):
+    return AD(sonic_vector=V_(**note_dict), **adsr_dict)
