+++ music/core/synths/CanonicalSynth.py
@@ -1,8 +1,6 @@
-# -*- coding: utf-8 -*-
 import numpy as n
 import music as M
-#from . import tables
-H = M.utils.H
+
 class CanonicalSynth:
     """Simple synth for sound synthesis with vibrato, tremolo and ADSR.
     
@@ -164,133 +162,3 @@ class CanonicalSynth:
         sound=self.table[Gamma_i%int(Lt)] # busca dos índices na tabela
         sound=self.adsrApply(sound)
         return sound
-
-def makeGaussianNoise(self,mean,std,DUR=2):
-    Lambda = DUR*self.samples_beat # Lambda sempre par
-    df = self.samplerate/float(Lambda)
-    MEAN=mean
-    STD=.1
-    coefs = n.exp(1j*n.random.uniform(0, 2*n.pi, Lambda))
-    # real par, imaginaria impar
-    coefs[Lambda/2+1:] = n.real(coefs[1:Lambda/2])[::-1] - 1j * \
-        n.imag(coefs[1:Lambda/2])[::-1]
-    coefs[0] = 0.  # sem bias
-    if Lambda%2==0:
-        coefs[Lambda/2] = 0.  # freq max eh real simplesmente
-
-    # as frequências relativas a cada coeficiente
-    # acima de Lambda/2 nao vale
-    fi = n.arange(coefs.shape[0])*df
-    f0 = 15.  # iniciamos o ruido em 15 Hz
-    f1=(mean-std/2)*3000
-    f2=(mean+std/2)*3000
-    i1 = n.floor(f1/df)  # primeiro coef a valer
-    i2 = n.floor(f2/df)  # ultimo coef a valer
-    coefs[:i1] = n.zeros(i1)
-    coefs[i2:]=n.zeros(len(coefs[i2:]))
-
-    # obtenção do ruído em suas amostras temporais
-    ruido = n.fft.ifft(coefs)
-    r = n.real(ruido)
-    r = ((r-r.min())/(r.max()-r.min()))*2-1
-
-    # fazer tre_freq variar conforme measures2
-    return r
-
-
-class IteratorSynth(CanonicalSynth):
-    """A synth that iterates through arbitrary lists of variables
-    
-
-    Any variable used by the CanonicalSynth can be used.
-    Just append the variable name with the token _sequence.
-
-    Example:
-    >>> isynth=M.IteratorSynth()
-    >>> isynth.fundamental_frequency_sequence = [220, 400, 100, 500]
-    >>> isynth.duration_sequence = [2, 1, 1.5]
-    >>> isynth.vibrato_frequency_sequence = [3, 6.5, 10]
-    >>> sounds=[]
-    >>> for i in range(300):
-            sounds += [isynth.renderIterate(tremolo_frequency=.2*i)]
-    >>> M.utils.write(M.H(*sounds),"./example.wav")
-
-    """
-
-    def renderIterate(self,**statevars):
-        self.absorbState(**statevars)
-        self.iterateElements()
-        return self.render()
-
-#     def render(self):
-#         sequences = [var for var in dir(self) if var.endswith("_sequence")]
-#         lens = [len(self.__dict__[seq]) for seq in sequences]
-#         iterations = max(lens)
-#         sonic_vector = [self.renderIterate() for i in range(iterations)]
-#         return sonic_vector
-
-    def iterateElements(self):
-        sequences=[var for var in dir(self) if var.endswith("_sequence")]
-        state_vars=[i[:-9] for i in sequences]
-        positions=[i+"_position" for i in sequences]
-        for sequence,state_var,position in zip(sequences,state_vars,positions):
-            if position not in dir(self):
-                self.__dict__[position]=0
-            self.__dict__[state_var]=self.__dict__[sequence][self.__dict__[position]]
-            self.__dict__[position]+=1
-            self.__dict__[position]%=len(self.__dict__[sequence])
-
-
-def sequenceOfStretches(x, s=[1,4,8,12], fs=44100):
-    """
-    Makes a sequence of squeezes of the fragment in x.
-
-    Parameters
-    ----------
-    x : array_like
-        The samples made to repeat as original or squeezed.
-        Assumed to be in the form (channels, samples),
-        i.e. x[1][120] is the 120th sample of the second channel.
-    s : list of numbers
-        Durations in seconds for each repeat of x.
-
-    Examples
-    --------
-    >>> asound = H(*[V(f=i, fv=j) for i, j in zip([220,440,330,440,330],
-                                                  [.5,15,6,5,30])])
-    >>> s = sequenceOfStretches(asound)
-    >>> s = sequenceOfStretches(asound,s=[.2,.3]*10+[.1,.2,.3,.4]*8+[.5,1.5,.5,1.,5.,.5,.25,.25,.5, 1., .5]*2)
-    >>> W(s, 'stretches.wav')
-    Notes
-    -----
-    This function is useful to render musical sequences given any material.
-
-    """
-    x = n.array(x)
-
-    s_ = s*fs
-    if len(x.shape) == 1:
-        l = x.shape[0]
-        stereo = False
-    else:
-        l = x.shape[1]
-        stereo = True
-    ns = l/fs
-    ns_ = [ns/i for i in s]
-    # x[::ns] (mono) or x[:, ::ns] stereo is the sound in one second
-    # for any duration s[i], use ns_ = ns//s[i]
-    # x[n.arange(0, len(x), ns_[i])]
-    sound = []
-    for ss in s:
-        if ns/ss >= 1:
-            indexes = n.arange(0, l, ns/ss).round().astype(n.int64)
-        else:
-            indexes = n.arange(0, l-1, ns/ss).round().astype(n.int64)
-        if stereo:
-            segment = x[:, indexes]
-        else:
-            segment = x[indexes]
-        sound.append(segment)
-    sound_ = H(*sound)
-    return sound_
-
