+++ music/core/synths/__init__.py
@@ -1,0 +1,21 @@
+from .CanonicalSynth import CanonicalSynth
+from .IteratorSynth import IteratorSynth
+
+from .d import D
+from .d_ import D_
+from .fm import FM
+from .loc import loc
+from .loc2 import loc2
+from .loc_ import loc_
+from .n import N
+from .n_ import N_
+from .noises import noises
+from .p import P
+from .pv import PV
+from .pv_ import PV_
+from .pvv import PVV
+from .t import T
+from .t_ import T_
+from .trill import trill
+from .v import V
+from .v_ import V_
